+++ lib/logging/benefits_intake/monitor.rb
@@ -73,7 +73,8 @@ module Logging
           claim:,
           user_account_uuid:,
           benefits_intake_uuid: lighthouse_service&.uuid,
-          message: e&.message
+          message: e&.message,
+          call_location: caller_locations.second
         )
       end
 
@@ -92,7 +93,8 @@ module Logging
           "#{submission_stats_key}.exhausted",
           claim: claim || msg['args'].first,
           user_account_uuid:,
-          message: msg
+          message: msg,
+          call_location: caller_locations.second
         )
 
         if claim
